:recycle: 지도 생성 로직 useTreeMap 커스텀 훅으로 분리
